+++ mysite/unmasque/src/pipeline/fragments/DisjunctionPipeLine.py
@@ -1,29 +1,28 @@
 import copy
 from abc import abstractmethod, ABC
 
-from ...src.core.aoa import AlgebraicPredicate
-from ...src.core.cs2 import Cs2
-from ...src.core.db_restorer import DbRestorer
-from ...src.core.equi_join import U2EquiJoin
-from ...src.core.filter import Filter
-from ...src.core.view_minimizer import ViewMinimizer
-from ...src.pipeline.abstract.generic_pipeline import GenericPipeLine
-from ...src.util.aoa_utils import get_constants_for
-from ...src.util.constants import FILTER, INEQUALITY, DONE, RUNNING, START, EQUALITY, DB_MINIMIZATION, \
+from mysite.unmasque.src.core.aoa import AlgebraicPredicate
+from mysite.unmasque.src.core.cs2 import Cs2
+from mysite.unmasque.src.core.db_restorer import DbRestorer
+from mysite.unmasque.src.core.equi_join import U2EquiJoin
+from mysite.unmasque.src.core.filter import Filter
+from mysite.unmasque.src.core.view_minimizer import ViewMinimizer
+from mysite.unmasque.src.pipeline.abstract.generic_pipeline import GenericPipeLine
+from mysite.unmasque.src.util.aoa_utils import get_constants_for
+from mysite.unmasque.src.util.constants import FILTER, INEQUALITY, DONE, RUNNING, START, EQUALITY, DB_MINIMIZATION, \
     SAMPLING, RESTORE_DB
-from ...src.util.utils import get_format, get_val_plus_delta
+from mysite.unmasque.src.util.utils import get_format, get_val_plus_delta
 
 
 class DisjunctionPipeLine(GenericPipeLine, ABC):
 
     def __init__(self, connectionHelper, name):
-        super().__init__(connectionHelper, name)
+        GenericPipeLine.__init__(self, connectionHelper, name)
         self.aoa = None
         self.equi_join = None
         self.filter_extractor = None
         self.db_restorer = None
         self.global_min_instance_dict = None
-
         self.key_lists = None
 
     def mutation_pipeline(self, core_relations, query, time_profile, restore_details=None):
